requirements.txt
streamlit
tensorflow
opencv-python
Pillow
numpy
matplotlib
scikit-learn
